test routes.json
