+++ web/src/services/produce.jsx
@@ -3,7 +3,7 @@ const host = import.meta.env.VITE_API_URL
 const baseUrl = `${host}/produce`
 
 const getAll = async () => {
-    console.log(`accessing ${baseUrl}`);
+    // console.log(`accessing ${baseUrl}`);
     const response = await axios.get(baseUrl)
     return response.data
 }
